Revert "LPD-51094 How can this ever be null?"
This reverts commit 6c02360e3a00a11352dc28b8144383e54828d392.

+++ modules/apps/sharing/sharing-test/src/testIntegration/java/com/liferay/sharing/internal/model/listener/test/GroupModelListenerTest.java
@@ -64,6 +64,7 @@ public class GroupModelListenerTest {
 	@Test
 	public void testDeletingGroupDeletesSharingEntries() throws Exception {
 		long classPK = _group.getGroupId();
+
 		UserGroup userGroup = UserGroupTestUtil.addUserGroup();
 
 		try {
@@ -100,7 +101,9 @@ public class GroupModelListenerTest {
 				groupSharingEntries.toString(), 0, groupSharingEntries.size());
 		}
 		finally {
-			_userGroupLocalService.deleteUserGroup(userGroup);
+			if (userGroup != null) {
+				_userGroupLocalService.deleteUserGroup(userGroup);
+			}
 		}
 	}
